# Add optimised canonical bitpacking for Aarch64. Closes #435.