… larq_compute_engine/core/macros.h
… larq_compute_engine/core/tests/packbits_aarch64_tests.cc
… larq_compute_engine/core/tests/packbits_tests.cc
… larq_compute_engine/tflite/kernels/bconv2d_impl.h
… larq_compute_engine/tflite/kernels/bmaxpool.cc
… larq_compute_engine/tflite/tests/bconv2d_int8_test.cc
… larq_compute_engine/tflite/tests/bconv2d_op_model.h
… larq_compute_engine/tflite/tests/bconv2d_test.cc
… larq_compute_engine/tflite/tests/bmaxpool_test.cc