## What do these changes do?

This PR removes the optimised non-canonical bitpacking for Aarch64 and replaces it with an optimised canonical bitpacking implementation instead. This means that all bitpacking is now canonical, as discussed in #435. As a result, the bitpacking code is a lot simpler, and a bunch of code templating is removed. 

## How has this been tested?

The 'big' kernel tests for Aarch64 (that don't run on CI) pass locally. The Aarch64 bitpacking tests have been strengthened to test for canonical bit ordering.

## Benchmark results

I have benchmarked current master (the 'Baseline') and this PR on my Android phone (a OnePlus 6). I used the LCE benchmark utility with `num_runs=250` and report the average and the standard-deviation.

#### 1 thread

| Model         |    Baseline   |       PR      |
|---------------|:-------------:|:-------------:|
| QuickNet      | 16.53 +- 0.06 | 16.53 +- 0.06 |
| QuickNetLarge | 24.21 +- 0.16 | 24.20 +- 0.08 |
| QuickNetXL    | 42.26 +- 0.28 | 42.09 +- 0.18 |

#### 4 threads

| Model         |    Baseline   |       PR      |
|---------------|:-------------:|:-------------:|
| QuickNet      | 6.92 +- 0.26  | 6.86 +- 0.54  |
| QuickNetLarge | 9.99 +- 0.23  | 10.01 +- 0.19 |
| QuickNetXL    | 17.01 +- 0.23 | 17.05 +- 0.33 |

These results show that there is no performance regression with this PR.

(I don't currently have my Raspberry Pi to hand so I can't run the benchmarks there, but when I did so for an earlier version of this PR I didn't see any significant difference from the baseline on that device, either.)

## Related issue number

Closes #435.