@Tombana 2020-07-31
Now that all the ordering is canonical, bitpacking into 2 `uint32`s is the same as bitpacking into 1 `uint64`. So if we have a matrix with 2 rows and 32 elements per row, we can just bitpack this as a single `uint64`.
I propose:
- the `packbits_matrix` function needs the `TBitpacked` type to distinguish block sizes 32 and 64 in order to properly do padding. It will then either call the `packbits_array` function many times (in case there's padding at each row) or just once: this is an optimization Adam already implemented here.
- However the `packbits_array` function (whether it was called on the flat array or on individual rows) should just ignore the `TBitpacked` type almost completely: it should just bitpack in whatever the most efficient way is, which might be blocks of size 128, even if `TBitpacked` was `uint8`. The only thing it needs `TBitpacked` for, is if there's padding, to know how much to pad.
- This could simplify a lot of code: instead of having those awfully long generic `pack_bitfield` functions, we could just have the 32-bit or 64-bit generic function and use only that one.

We can do this either in this PR or in a separate one, but I think it would be good to do. What do you think @AdamHillier ?