@AdamHillier 2020-07-31
> We can do this either in this PR or in a separate one, but I think it would be good to do. What do you think @AdamHillier ?
> 
> EDIT: I just saw #446, that makes it even better: we can remove `TBitpacked` completely (and still use any blocksize like 64 (or 128 if we want) in the `packbits_array` part).

Haha yeah your comment here reminded me that I needed to write-up #446. But yeah, I definitely agree that there's more to optimise here, but I think we can do that as part of the process of standardising 32-bit bitpacking.